api/requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
moviepy
edge-tts
openai
python-multipart
python-dotenv
requests
lxml
aiofiles
